refactor: reduce dom.js hub coupling by routing imports through domain facades
Redirect 14 modules that imported directly from dom.js to use the
appropriate domain facade instead (terminal-dom, file-dom, tab-dom,
workspace-dom, chevron-row, or the new git-dom). This drops dom.js
fan-in from 25 direct importers to 11 (8 facade re-exporters + 3 core
utilities).

Changes:
- Create git-dom.js facade for worktree-flow, worktree-dialog, open-pr-flow
- Add renderButtonBar to terminal-dom.js for board-view
- Redirect file-domain modules to file-dom.js
- Redirect buildChevronRow consumers to chevron-row.js
- Redirect workspace-level views to workspace-dom.js
