+++ src/utils/worktree-flow.js
@@ -9,7 +9,7 @@
 
 import { showWorktreeDialog } from './worktree-dialog.js';
 import { showConfirmDialog, showErrorAlert } from './dom-dialogs.js';
-import { _el } from './dom.js';
+import { _el } from './git-dom.js';
 
 /**
  * Branches to hide from the "existing branch" picker: those already checked
